fix: reach uv and the bun debugger with the forwarded network settings
