+++ debugger/Dockerfile
@@ -46,6 +46,7 @@ WORKDIR /app
 # Copy the debug service files
 COPY dap_debug_service.ts .
 COPY dap_websocket_server_bun.ts .
+COPY env_passthrough.ts .
 COPY dap_websocket_server.py .
 
 # Expose the default port
