+++ docker/DockerfileExtra
@@ -94,6 +94,7 @@ WORKDIR /debugger
 # Copy debugger files
 COPY debugger/dap_debug_service.ts .
 COPY debugger/dap_websocket_server_bun.ts .
+COPY debugger/env_passthrough.ts .
 COPY debugger/dap_websocket_server.py .
 COPY debugger/nsjail.debug.config.proto .
